Move FlatRiemannianMetric to its own file
