+++ geomstats/learning/frechet_mean.py
@@ -12,7 +12,7 @@ from sklearn.base import BaseEstimator
 import geomstats.backend as gs
 import geomstats.errors as error
 from geomstats.geometry.discrete_curves import ElasticMetric, SRVMetric
-from geomstats.geometry.euclidean import FlatRiemannianMetric
+from geomstats.geometry.flat_riemannian_metric import FlatRiemannianMetric
 from geomstats.geometry.hypersphere import HypersphereMetric
 
 ELASTIC_METRICS = [SRVMetric, ElasticMetric]
