+++ tests/tests_geomstats/test_geometry/test_euclidean.py
@@ -2,37 +2,23 @@ import random
 
 import pytest
 
-from geomstats.geometry.euclidean import Euclidean, FlatRiemannianMetric
-from geomstats.geometry.spd_matrices import SPDMatrices
+from geomstats.geometry.euclidean import Euclidean
 from geomstats.test.parametrizers import DataBasedParametrizer
 from geomstats.test_cases.geometry.euclidean import (
     EuclideanMetricTestCase,
     EuclideanTestCase,
-    FlatRiemannianMetricTestCase,
 )
 
 from .data.euclidean import (
     EuclideanMetric2TestData,
     EuclideanMetricTestData,
     EuclideanTestData,
-    FlatRiemannianMetricTestData,
 )
 
 
 class TestEuclidean(EuclideanTestCase, metaclass=DataBasedParametrizer):
     space = Euclidean(dim=random.randint(2, 5), equip=False)
     testing_data = EuclideanTestData()
-
-
-class TestFlatRiemannianMetric(
-    FlatRiemannianMetricTestCase, metaclass=DataBasedParametrizer
-):
-    _dim = random.randint(2, 5)
-    space = Euclidean(_dim, equip=False).equip_with_metric(
-        FlatRiemannianMetric,
-        metric_matrix=SPDMatrices(_dim, equip=False).random_point(),
-    )
-    testing_data = FlatRiemannianMetricTestData()
 
 
 class TestEuclideanMetric(EuclideanMetricTestCase, metaclass=DataBasedParametrizer):
